openclaw-adapter/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio==5.11.4
python-multipart==0.0.9
pymongo==4.8.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
requests==2.32.3
